feat(web): decimal-aware money guards for onboarding budget + split
Budget step is now free-entry min/max PHP fields (was preset bands) with
end-to-end guards:
- input: sanitize on keystroke (digits + one dot, max 2 decimals), inputMode
  decimal; parseMoney strips currency/commas, rejects negatives, clamps to a
  100M ceiling, rounds to centavos, never NaN.
- validation: min <= max enforced in the UI (inline error + blocked continue/
  create) and in /api/family/create (<=2dp, >=0, <=ceiling).
- storage: budget_min/budget_max migration changed integer -> numeric(12,2)
  so centavos persist.

SplitEditor: empty/NaN field now coerces to 0 before the 0..100 clamp, fixing
a gap where clearing a percent field poisoned the split sum (helps the
dashboard settings too).

+++ docs/sql/2026-07-04-family-wallets-onboarding-fields.sql
@@ -11,11 +11,13 @@
 --   percents (NOT NULL), policy_json (NOT NULL), savings_lock_all_admins (NOT NULL)
 -- None of the new column names collided.
 
+-- Budget bounds are numeric(12,2): PHP with centavos, up to 10 integer digits
+-- (ceiling well above the app's 100M guard) and 2 decimal places.
 alter table public.family_wallets
   add column if not exists household_type text
     check (household_type in ('family-at-home', 'both-abroad', 'scratch')),
-  add column if not exists budget_min integer check (budget_min >= 0),
-  add column if not exists budget_max integer check (budget_max >= 0);
+  add column if not exists budget_min numeric(12, 2) check (budget_min >= 0),
+  add column if not exists budget_max numeric(12, 2) check (budget_max >= 0);
 
 -- Range sanity: if both bounds are present, min <= max. Nullable-friendly.
 -- Guarded so re-running the script doesn't error on the existing constraint.
